+++ forms.py
@@ -15,7 +15,7 @@ class FormAddCarga(FlaskForm):
 
 class FormInfo(FlaskForm):
     dias = IntegerField('Nº de dias úteis', validators= [DataRequired(),NumberRange(min=1,max=31,message="Valor inválido")],default=22)
-    ponta = IntegerField('Início do horário de ponta', validators= [DataRequired(),NumberRange(min=0,max=21,message="Valor inválido")],default=18)
+    ponta = FloatField('Início do horário de ponta', validators= [DataRequired(),NumberRange(min=0,max=21,message="Valor inválido")],default=18)
 
     registrar_info = SubmitField('Registrar', validators= [DataRequired()])
